msm8974: Increase boost frequencies for non-primary CPUs
Gaming (especially 2D) took a hit with the new Interactive configuration.

Increase the boost frequencies for non-primary CPUs to improve responsiveness when more CPUs are online (and thus when load is greater). Non-primary CPUs are boosted for very short amounts of time, so this shouldn't impact battery life too badly.

Signed-off-by: Louis Popi <theh2o64@gmail.com>